Add buttons for Main search
